# app/govern/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bll.GetPicCaptchaResp:
    properties:
      picCaptchaB64s:
        type: string
      picCaptchaId:
        type: string
    type: object
...
  bll.ListStaffResp:
    properties:
      data:
        type: object
      total:
        type: integer
    type: object
...
  bll.ProfileStaffResp:
    properties:
      account:
        type: string
      email:
        type: string
      enable:
        type: boolean
      gender:
        type: string
      lastSignInIp:
        type: string
      lastSignInTime:
        type: integer
      phone:
        type: string
      realName:
        type: string
      roles:
        items:
          type: string
        type: array
      signInIpWhiteList:
        items:
          type: string
        type: array
      signUpAt:
        type: integer
      updatedAt:
        type: integer
    type: object
...
  bll.SignInReq:
    properties:
      account:
        type: string
      password:
        type: string
      picCaptchaAnswer:
        type: string
      picCaptchaId:
        type: string
      timestamp:
        type: integer
    required:
    - picCaptchaAnswer
    - picCaptchaId
    type: object
  bll.SignInResp:
    properties:
      expiresAt:
        type: integer
      token:
        type: string
    type: object
  bll.SignUpReq:
    properties:
      account:
        type: string
      email:
        type: string
      gender:
        type: string
      password:
        type: string
      phone:
        type: string
      realName:
        type: string
      signInIpWhitelist:
        items:
          type: string
        type: array
      timestamp:
        type: integer
    required:
    - email
    - timestamp
    type: object
...
  bll.UpdateStaffEmailReq:
    properties:
      captchaFromOldEmail:
        type: string
      newEmail:
        type: string
    required:
    - newEmail
    type: object
  bll.UpdateStaffPasswordReq:
    properties:
      newPassword:
        type: string
      newPasswordRepeat:
        type: string
      oldPassword:
        type: string
    type: object
  bll.UpdateStaffSignInIpWhitelistReq:
    properties:
      signInIpWhitelist:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8000
info:
  contact:
    email: sfshf@github.com
    name: gavin
    url: http://github.com/sfshf
  description: This is a back-end government app.
  license:
    name: MIT
    url: https://github.com/sfshf/sprout/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Govern APIs
  version: 0.0.1-beta
paths:
...
  /picCaptcha:
    get:
      description: Get a new picture captcha id and base64 string of the picture,
        and delete the obsolete captcha of the obsolete_id, if has.
      operationId: get-pic-captcha
      parameters:
      - description: an obsolete captcha id.
        in: query
        name: obsolete_id
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: captcha id and picture.
          schema:
            $ref: '#/definitions/bll.GetPicCaptchaResp'
      summary: Get a picture captcha.
      tags:
      - staff
  /picCaptchaAnswer:
    get:
      description: Get the answer code of a picture captcha with the captcha id.
      operationId: get-pic-captcha-answer
      parameters:
      - description: a captcha id.
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: Invalid Token, Invalid Captcha, Unauthorized, or other errors.
          schema:
            type: error
        "2000":
          description: captcha answer code
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get the answer code of a picture captcha.
      tags:
      - staff
...
  /signIn:
    patch:
      consumes:
      - application/json
      description: Sign in with account and password, supporting picture captcha authentication.
      operationId: sign-in
      parameters:
      - description: required attributes to sign in.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.SignInReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: sign in token and expiry time.
          schema:
            $ref: '#/definitions/bll.SignInResp'
      summary: Sign in.
      tags:
      - staff
  /signOff/:id:
    delete:
      description: Sign off the session account, or some specific account only by
        root account.
      operationId: sign-off
      parameters:
      - description: id of the account to sign off.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Sign off.
      tags:
      - staff
  /signOut:
    patch:
      description: Sign out the session account.
      operationId: sign-out
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Sign out.
      tags:
      - staff
  /signUp:
    post:
      consumes:
      - application/json
      description: Sign up a new staff account.
      operationId: sign-up
      parameters:
      - description: required attributes to sign up.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.SignUpReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      summary: Sign up.
      tags:
      - staff
  /staffs:
    get:
      description: Get a list of staff accounts.
      operationId: staff-list
      parameters:
      - in: query
        name: account
        type: string
      - in: query
        name: email
        type: string
      - in: query
        name: enable
        type: boolean
      - in: query
        name: gender
        type: string
      - in: query
        name: lastSignInIp
        type: string
      - in: query
        name: lastSignInTimeBegin
        type: integer
      - in: query
        name: lastSignInTimeEnd
        type: integer
      - in: query
        name: orderBy
        type: string
      - in: query
        name: page
        type: integer
      - in: query
        name: perPage
        type: integer
      - in: query
        name: phone
        type: string
      - in: query
        name: role
        type: string
      - in: query
        name: signIn
        type: boolean
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: staff list.
          schema:
            $ref: '#/definitions/bll.ListStaffResp'
      security:
      - ApiKeyAuth: []
      summary: Get a list of staff accounts.
      tags:
      - staff
  /staffs/:id:
    get:
      description: Get the profile of a staff account.
      operationId: staff-profile
      parameters:
      - description: id of the staff account.
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: profile of the staff.
          schema:
            $ref: '#/definitions/bll.ProfileStaffResp'
      security:
      - ApiKeyAuth: []
      summary: Get infos of a staff account.
      tags:
      - staff
  /staffs/:id/email:
    patch:
      consumes:
      - application/json
      description: Update the email of a staff account.
      operationId: staff-update-email
      parameters:
      - description: id of the staff account to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update the email.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateStaffEmailReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update the email of a staff account.
      tags:
      - staff
  /staffs/:id/enable:
    patch:
      consumes:
      - application/json
      description: Enable or disable a staff account.
      operationId: staff-enable
      parameters:
      - description: id of the staff account.
        in: path
        name: id
        required: true
        type: string
      - description: true for enable, or false for disable.
        in: body
        name: enable
        required: true
        schema:
          type: boolean
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Enable or disable a staff account.
      tags:
      - staff
  /staffs/:id/password:
    patch:
      consumes:
      - application/json
      description: Update the password of a staff account.
      operationId: staff-update-password
      parameters:
      - description: id of the staff account to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update the password.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateStaffPasswordReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update the password of a staff account.
      tags:
      - staff
  /staffs/:id/phone:
    patch:
      consumes:
      - application/json
      description: Update the phone of a staff account.
      operationId: staff-update-phone
      parameters:
      - description: id of the staff account to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update the phone.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateStaffPasswordReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update the phone of a staff account.
      tags:
      - staff
  /staffs/:id/roles:
    patch:
      consumes:
      - application/json
      description: Update the roles of a staff account.
      operationId: staff-update-roles
      parameters:
      - description: id of the staff account to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update the roles.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateStaffPasswordReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update the roles of a staff account.
      tags:
      - staff
  /staffs/:id/signInIpWhitelist:
    patch:
      consumes:
      - application/json
      description: Update the sign-in ip-whitelist of a staff account.
      operationId: staff-update-signInIpWhitelist
      parameters:
      - description: id of the staff account to update.
        in: path
        name: id
        required: true
        type: string
      - description: attributes need to update the sign-in ip-whitelist.
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/bll.UpdateStaffSignInIpWhitelistReq'
      produces:
      - application/json
      responses:
        "1000":
          description: feasible and predictable errors.
          schema:
            type: error
        "2000":
          description: successful action.
          schema:
            type: "null"
      security:
      - ApiKeyAuth: []
      summary: Update the sign-in ip-whitelist of a staff account.
      tags:
      - staff
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"